+++ src/react/components/autocompleter/XUIAutocompleter.js
@@ -39,6 +39,16 @@ function getHandlers(instance) {
 				if (instance.ddt.isDropDownOpen()) {
 					instance.dropdown.onKeyDown(event);
 				}
+
+				if (
+					event.key === 'Backspace' &&
+					instance.input.inputNode.value === "" &&
+					instance.props.onBackspacePill &&
+					instance.props.pills &&
+					(instance.props.pills.length > 0 || React.isValidElement(instance.props.pills))
+				) {
+					instance.props.onBackspacePill();
+				}
 			},
 			onInputFocus: () => {
 				if (!instance.state.focused) {
@@ -102,7 +112,9 @@ export default class XUIAutocompleter extends PureComponent {
 	}
 
 	scrollInputIntoView() {
-		this.trigger.scrollLeft = this.trigger.scrollWidth;
+		if(this.trigger !== null){
+			this.trigger.scrollLeft = this.trigger.scrollWidth;
+		}
 	}
 
 	/**
@@ -154,10 +166,12 @@ export default class XUIAutocompleter extends PureComponent {
 		let inputQaHook = null;
 		let listQaHook = null;
 		let containerQaHook = null;
+		let dropdownQaHook = null;
 		if (props.qaHook) {
 			inputQaHook = `${props.qaHook}--input`;
 			listQaHook = `${props.qaHook}--list`;
 			containerQaHook = `${props.qaHook}--container`;
+			dropdownQaHook = `${props.qaHook}--dropdown`;
 		}
 
 		const handlers = getHandlers(completer);
@@ -202,7 +216,7 @@ export default class XUIAutocompleter extends PureComponent {
 			<DropDown
 				ref={c => completer.dropdown = c}
 				ignoreKeyboardEvents={ignoreKeyboardEvents}
-				id={props.id}
+				id={props.dropdownId}
 				onSelect={props.onOptionSelect}
 				hasKeyboardEvents={false}
 				className={props.dropdownClassName}
@@ -231,6 +245,7 @@ export default class XUIAutocompleter extends PureComponent {
 				onFocus={handlers.onFocus}
 				onBlur={handlers.onBlur}
 				data-automationid={containerQaHook}
+				id={props.id}
 			>
 				<DropDownToggled
 					ref={c => completer.ddt = c}
@@ -242,6 +257,7 @@ export default class XUIAutocompleter extends PureComponent {
 					triggerClickAction="none"
 					forceDesktop={props.forceDesktop}
 					matchTriggerWidth={props.matchTriggerWidth && !props.dropdownSize}
+					qaHook={dropdownQaHook}
 				/>
 			</div>
 		);
@@ -252,13 +268,19 @@ XUIAutocompleter.propTypes = {
 	/** Callback to handle when an option has been selected from the dropdown */
 	onOptionSelect: PropTypes.func,
 
+	/** Callback to handle when a pill has been backspaced */
+	onBackspacePill: PropTypes.func,
+
 	/** When set to true a loader will be displayed instead of the picklist items.
 	 * State for this should be managed externally and it's defaulted to false.
 	 */
 	loading: PropTypes.bool,
 
-	/** ID to be added to the dropdown element of the completer */
+	/** ID to be added to the root node of the completer */
 	id: PropTypes.string,
+
+	/** ID to be added to the dropdown element of the completer */
+	dropdownId: PropTypes.string,
 
 	/** Value that should be inside the input. */
 	searchValue: PropTypes.string,