+++ src/react/components/dropdown/__tests__/DropDownFooter-test.js
@@ -3,7 +3,7 @@ import renderer from 'react-test-renderer';
 import DropDownFooter from '../DropDownFooter.js';
 
 describe('<DropDownFooter />', () => {
-    it ('shoudl render an auotmation id when a qaHook is passed', () => {
+    it ('should render an auotmation id when a qaHook is passed', () => {
         const automationId = renderer.create(
             <DropDownFooter qaHook="dropdownfooter-example">content</DropDownFooter>
         );