+++ src/react/components/icon/__tests__/XUIIcon-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Enzyme, { mount, render } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-15';
+import Adapter from 'enzyme-adapter-react-16';
 import XUIIcon from '../XUIIcon';
 import accessibility from '@xero/xui-icon/icons/accessibility';