+++ src/react/components/input/__tests__/XUIInput-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Enzyme, { mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-15';
+import Adapter from 'enzyme-adapter-react-16';
 import renderer from 'react-test-renderer';
 import XUIInput from '../XUIInput';
 import accessibility from '@xero/xui-icon/icons/accessibility';